# web/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Mamao.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:33191"
    }
  ],
  "paths": {
...
    "/api/v1/departments": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "listDepartments",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DepartmentNode"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organization"
        ],
        "operationId": "createDepartment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDepartmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartmentNode"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/departments/{id}": {
      "put": {
        "tags": [
          "Organization"
        ],
        "operationId": "updateDepartment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDepartmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Organization"
        ],
        "operationId": "deleteDepartment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/departments/{id}/move": {
      "post": {
        "tags": [
          "Organization"
        ],
        "operationId": "moveDepartment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveDepartmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/positions": {
      "get": {
        "tags": [
          "Organization"
        ],
        "operationId": "listPositions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Organization"
        ],
        "operationId": "createPosition",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePositionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/positions/{id}": {
      "put": {
        "tags": [
          "Organization"
        ],
        "operationId": "updatePosition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePositionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Organization"
        ],
        "operationId": "deletePosition",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateDepartmentRequest": {
        "required": [
          "name",
          "parentId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "string",
            "oneOf": [
              {
                "$ref": "#/components/schemas/DepartmentId"
              }
            ],
            "nullable": true
          }
        }
      },
...
      "CreatePositionRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "DepartmentId": {
        "type": "string",
        "format": "uuid"
      },
      "DepartmentNode": {
        "required": [
          "id",
          "name",
          "parentId",
          "depth",
          "managerId",
          "managerName",
          "employeeCount",
          "subtreeEmployeeCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/DepartmentId"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": "string",
            "oneOf": [
              {
                "$ref": "#/components/schemas/DepartmentId"
              }
            ],
            "nullable": true
          },
          "depth": {
            "format": "int32",
            "type": "integer"
          },
          "managerId": {
            "type": "string",
            "oneOf": [
              {
                "$ref": "#/components/schemas/EmployeeId"
              }
            ],
            "nullable": true
          },
          "managerName": {
            "type": "string",
            "nullable": true
          },
          "employeeCount": {
            "format": "int32",
            "type": "integer"
          },
          "subtreeEmployeeCount": {
            "format": "int32",
            "type": "integer"
          }
        }
      },
      "EmployeeId": {
        "type": "string",
        "format": "uuid"
      },
...
      "MoveDepartmentRequest": {
        "required": [
          "parentId"
        ],
        "type": "object",
        "properties": {
          "parentId": {
            "type": "string",
            "oneOf": [
              {
                "$ref": "#/components/schemas/DepartmentId"
              }
            ],
            "nullable": true
          }
        }
      },
...
      "PositionId": {
        "type": "string",
        "format": "uuid"
      },
      "PositionResponse": {
        "required": [
          "id",
          "name",
          "employeeCount"
        ],
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/PositionId"
          },
          "name": {
            "type": "string"
          },
          "employeeCount": {
            "format": "int32",
            "type": "integer"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "format": "int32",
            "nullable": true,
            "type": "integer"
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "UpdateDepartmentRequest": {
        "required": [
          "name",
          "managerId"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "managerId": {
            "type": "string",
            "oneOf": [
              {
                "$ref": "#/components/schemas/EmployeeId"
              }
            ],
            "nullable": true
          }
        }
      },
...
      "UpdatePositionRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      }
    }
  },
...
}